% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeSamplingSequence.R
\name{cumulativeSamplingSequence}
\alias{cumulativeSamplingSequence}
\title{Cumulative Sampling Sequence}
\usage{
cumulativeSamplingSequence(N_1, T)
}
\arguments{
\item{N_1}{initial sample size}

\item{T}{number of sequential samples}
}
\value{
N_T sampling sequence
}
\description{
This function calculates the sample size of sequential samples for the equivalency experiment. Sample sizes are given in terms of cumulative sample size.
}
