% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStability.R
\name{processStability}
\alias{processStability}
\title{Check Process Stability}
\usage{
processStability(data, show.plot = TRUE)
}
\arguments{
\item{data}{values derived from chronological experiment}
}
\value{
The number of out of control points and x-bar chart
}
\description{
This function generates an x-bar chart of the variable measured from the reference process.
It also counts the number of points that fall outside of the control limits.
Data must be in the following form to work: it must have n rows (number of observations per subgroup) and m columns (number of subgroups) where the columns are in chronological order of measurement. Each cell must contain a numeric value.
}
