% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExperiment.R
\name{simulateExperiment}
\alias{simulateExperiment}
\title{Experimental Simulations}
\usage{
simulateExperiment(
  b = 8,
  alpha = 0.01,
  n_max = 499,
  simulations = 100,
  iterations = 100,
  TVecAcceptable,
  n1VecAcceptable
)
}
\arguments{
\item{b}{the number of bins used to characterize the reference distribution}

\item{alpha}{the significance level}

\item{n_max}{the highest possible sample size}

\item{simulations}{the number values of rho tested in simulations to find optimal sampling strategy}

\item{iterations}{the number of simulations per parameter value rho}

\item{TVecAcceptable}{vector containing values of T for possible experiments}

\item{n1VecAcceptable}{vector containing values of N_1 for possible experiments}
}
\value{
A dataframe containing the power levels, sample sizes, and false positive rates of various combinations of N_1 and T
}
\description{
This function simulates experiments where the null hypothesis is either true or not true where experimental designs set by the function possibleSamplingStrats,
}
