% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingRule.R
\name{stoppingRule}
\alias{stoppingRule}
\title{When to Stop Sampling}
\usage{
stoppingRule(cramersV, N, b, alpha = 0.01)
}
\arguments{
\item{cramersV}{measurment of how closely the candidate process resembles the reference process}

\item{N}{sample size corresponding to this value of cramers's v}

\item{b}{the number of bins}

\item{alpha}{significance level}
}
\value{
decision as to whether or not to continue sampling. If continueSampling == TRUE, keep going unless t = T. If continueSampling == FALSE, stop sampling
}
\description{
This function takes the cramers v of a sample and other aspects of the design to determine whether or not the statistic is outside the limits expected for that design given the null hypothesis that the candidate and reference processes are equivalent.
}
