% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateChiSquared.R
\name{candidateChiSquared}
\alias{candidateChiSquared}
\title{Candidate Process Chi Squared Value}
\usage{
candidateChiSquared(data, bins, show.plot = TRUE)
}
\arguments{
\item{data}{values derived from chronological experiment}

\item{bins}{percentiles of reference distribution}
}
\value{
chi squared value
}
\description{
This function calculates the chi squared value of the candidate process as it fits to the bins defined by the reference process
Data must be in the following form to work: it must have n rows (number of observations per subgroup) and m columns (number of subgroups) where the columns are in chronological order of measurement. Each cell must contain a numeric value. bins must be defined by quantile function (see tutorial)
}
