% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidateCramersV.R
\name{candidateCramersV}
\alias{candidateCramersV}
\title{Candidate Cramer's V}
\usage{
candidateCramersV(chiSquare, b, N)
}
\arguments{
\item{chiSquare}{a numeric value giving the goodness of fit of the candidate samples on the reference bins}

\item{b}{a numeric value indicating the number of bins}
}
\value{
\describe{
\item{N}{A numeric value indicating the sample size.}
}
}
\description{
Calculates the normalized Cramer's V of the candidate process based on the limits of the 
