% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeSamplingSequence.R
\name{cumulativeSamplingSequence}
\alias{cumulativeSamplingSequence}
\title{Cumulative Sampling Sequence}
\usage{
cumulativeSamplingSequence(N_1, T)
}
\arguments{
\item{N_1}{a numeric value indicating the initial sample size.}

\item{T}{a numeric value number of sequential samples.}
}
\value{
\describe{
\item{N_T}{A numeric vector giving the sample size sequence.}
}
}
\description{
Calculates the sample size of sequential samples for the equivalency experiment. Sample sizes are given in terms of cumulative sample size.
}
