% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalStrategy.R
\name{optimalStrategy}
\alias{optimalStrategy}
\title{Finding Optimal Sampling Strategy}
\usage{
optimalStrategy(
  data,
  wVec,
  L1 = 0,
  T1 = 1,
  T2 = 1,
  U2 = 41,
  T3 = 0,
  U3 = 1,
  k = 2
)
}
\arguments{
\item{data}{dataframe derived from simulateExperiment function}

\item{wVec}{a 3 element vector which represents the weights for each desirability function. The first index corresponds to power, the second to sample size, and the third to the false positive rate. Each value must be between 0 and 1 and the sum of all 3 must equal 1.}
}
\value{
Optimal sampling strategy given by N_1 and T, the final power level, expected sample size (given the two processes are not equivalent), and false positive rate of this design
}
\description{
This function runs simulations which finds the optimal sampling strategy given various experimental constraints.
}
